more work on routng in APPCS

+++ appCentralService/app.js
@@ -3,6 +3,7 @@ const reload = require('reload');
 const app = express();
 const io = require('socket.io')();
 const db = require('./db/db');
+const uuid = require('uuid');
 const EVENTS = require('./constants/socketEvents');
 const MAIN_NAMESPACE = '/main';
 
@@ -36,7 +37,7 @@ app.get('/appcs/game', (req, res) => {
         res.json({
             success: true,
             games
-        })
+        });
     }).catch((e) => {
         res.json({
             success: false,
@@ -49,16 +50,22 @@ app.get('/appcs/game', (req, res) => {
 AppCS Route.
  POST /appcs/game/create/:gameId : Create game
 
+POST body:
+{
+    game: {
+        // game object.
+    }
+}
  */
 app.post('/appcs/game/create', (req, res) => {
-    db.createGame(req.body.game).then(() => {
+    db.createGame(req.body.game).then((newgame) => {
+        // link used to go to the lobby page.
         io.of(MAIN_NAMESPACE).emit(EVENTS.GAME_CREATED, {
-            lobbylink: `/game/lobby/${req.params.gameId}`,
-            game: req.body.game
+            game: newgame
         });
         res.json({
             success: true,
-            lobbylink: `/game/lobby/${req.params.gameId}`
+            game: newgame
         })
     }).catch((e) => {
         res.json({
@@ -73,21 +80,33 @@ app.post('/appcs/game/create', (req, res) => {
 AppCS Route.
 DELETE /appcs/game/delete/:gameId: Delete game
 
+POST body:
+   {
+    socketId: "awdijaos2123eda",
+    gameId: "120d=qnsk=qo2n"
+   }
  */
 app.delete('/appcs/game/delete/:gameId', (req, res) => {
-    db.deleteGame(req.params.gameId).then(() => {
-        io.of(MAIN_NAMESPACE).emit(EVENTS.GAME_DELETED, {
-            gameId: req.body.gameId
-        });
-        io.of(MAIN_NAMESPACE).to(req.params.gameId).emit(EVENTS.LOBBY.LOBBY_GAME_DELETED);
-        res.json({
-            success: true,
-        })
-    }).catch((e) => {
-        res.json({
-            success: false,
-        })
+    let creator = undefined;
+    db.getGame(req.params.gameId).then((game) => {
+        creator = game.creator.socketId;
     });
+    if (creator === req.body.socketId)
+        db.deleteGame(req.params.gameId).then(() => {
+            io.of(MAIN_NAMESPACE).emit(EVENTS.GAME_DELETED, {
+                gameId: req.body.gameId
+            });
+            io.of(MAIN_NAMESPACE)
+                .to(req.params.gameId)
+                .emit(EVENTS.LOBBY.LOBBY_GAME_DELETED);
+            res.json({
+                success: true,
+            })
+        }).catch((e) => {
+            res.json({
+                success: false,
+            })
+        });
 });
 
 
@@ -96,31 +115,34 @@ app.delete('/appcs/game/delete/:gameId', (req, res) => {
 AppCS Route.
 POST /appcs/game/start/:gameId: Start game
 NOTE: we communicate with our GMS here.
+POST body:
+{
+    gameId:"awdnsoawd", // game id of one to be deleted.
+    socketId:"ajwodmala" // socket id of requester
+}
  */
 
-app.post('/appcs/game/start/:gameId',(req,res)=>{
+app.post('/appcs/game/start/:gameId', (req, res) => {
     let creatorId = undefined;
-    db.getGame(req.params.gameId).then((game)=>{
+    db.getGame(req.params.gameId).then((game) => {
         creatorId = game.creator.socketId;
     })
 
-    let supposedCreatorId = req.params.user.socketId; //authenticate.
-    if(creatorId === supposedCreatorId) {
+    let supposedCreatorId = req.params.socketId; //authenticate.
+    if (creatorId === supposedCreatorId) {
         io.of(MAIN_NAMESPACE).to(req.params.gameId)
         //TODO: communicate with GMS.
     }
     else
         res.json({
-            success:false
+            success: false
         })
 })
 
 
-
-
 io.of(MAIN_NAMESPACE).on('connect', (socket) => {
     if (!socket.sentMydata) {
-        socket.emit(EVENTS.CONN_DETAILS, {socketId: socket.id });
+        socket.emit(EVENTS.CONN_DETAILS, {socketId: socket.id});
         socket.sentMydata = true;
     }
 
@@ -131,12 +153,14 @@ io.of(MAIN_NAMESPACE).on('connect', (socket) => {
 
      */
     socket.on(EVENTS.LOBBY.CLIENT_ATTEMPT_JOIN, (clientUserObj, gameId) => {
-        let roomName = gameId;
+        let roomName = gameId; //TODO: roomName should be secret here.
+        // Could be a hash.
+
         db.joinGame(clientUserObj, gameId).then(() => {
             socket.join(roomName);
             io.of(MAIN_NAMESPACE)
                 .to(`${gameId}`)
-                .emit('userJoined', {...clientUserObj,socketId:null});
+                .emit('userJoined', {...clientUserObj, socketId: null});
             socket.emit(EVENTS.LOBBY.CLIENT_ATTEMPT_JOIN_ACK);
         }).catch((e) => {
             socket.emit(EVENTS.LOBBY.CLIENT_ATTEMPT_JOIN_NOACK);
@@ -154,7 +178,7 @@ io.of(MAIN_NAMESPACE).on('connect', (socket) => {
             //note: .to() is the one for room in the main namespace.
             let joinedRoom = clientUserObj.joinedRoom;
             socket.leave(`${joinedRoom}`);
-            io.of(MAIN_NAMESPACE).to(`${joinedRoom}`).emit('userLeft', {...clientUserObj,socketId:null});
+            io.of(MAIN_NAMESPACE).to(`${joinedRoom}`).emit('userLeft', {...clientUserObj, socketId: null});
             socket.emit(EVENTS.LOBBY.CLIENT_LEAVE_ACK);
         }).catch((e) => {
             socket.emit(EVENTS.LOBBY.CLIENT_LEAVE_NOACK);
